+++ vllm/model_executor/layers/attention/flash.py
@@ -5,10 +5,9 @@ from flash_attn import flash_attn_func
 import torch
 
 from vllm.model_executor.input_metadata import InputMetadata
-from vllm.model_executor.layers.attention_backends.base import BaseAttention
-from vllm.model_executor.layers.attention_backends.paged_attn import (
-    PagedAttention)
-from vllm.model_executor.layers.attention_backends.utils import expand_gqa
+from vllm.model_executor.layers.attention.base import BaseAttention
+from vllm.model_executor.layers.attention.paged_attn import PagedAttention
+from vllm.model_executor.layers.attention.utils import expand_gqa
 
 
 class Attention(BaseAttention):
